requirements.txt
pyttsx3==2.90
flask==2.3.3
pyyaml==6.0.1
pygame==2.5.2